24.03.05 | 이지후 | 코스 변경 및 aicc 작업중 1박 2일까지 출력 완료 코스 저장 작업 진행 해야함
