+++ src/main/resources/templates/nav/nav.html
@@ -47,9 +47,9 @@
 						<li class="nav-item"><a class="nav-link" href="band">밴드</a></li>
 						<li class="nav-item dropdown navli"><a class="nav-link" href="info"
 							role="button" data-bs-toggle="dropdown" aria-expanded="false">여행정보</a>
-							<ul class="dropdown-menu z-5 text-center" >
+							<ul class="dropdown-menu z-5 text-md-center w-25" >
 								<li ><a class="dropdown-item " href="info">여행정보</a></li>
-								<li><a class="dropdown-item rgyPostIt" href="infoCourse">여행코스</a></li>
+								<li><a class="dropdown-item " href="infoCourse">여행코스</a></li>
 							</ul></li>
 						<li class="nav-item"><a class="nav-link mb-3 mb-md-0"
 							href="map">지도</a></li>
