+++ src/main/resources/templates/thymeleaf/Member/mypage.html
@@ -253,7 +253,7 @@
 									<th:block th:if="${#lists.size(course) < 5}">
 										<a
 											th:href="@{'mypageCourseAll?id='+${loginMember.getMember_Id()}}"
-											class="gg-dim-txt f13">전체보기</a>
+											class="gg-dim-txt f13">내 코스 전체 보기</a>
 									</th:block>
 									
 								</div>
