+++ src/main/resources/templates/thymeleaf/aicc/aicc.html
@@ -30,8 +30,8 @@
                data-bs-slide-to="1" aria-label="Slide 2"></button>
             <button type="button" data-bs-target="#ai-cok-carousel"
                data-bs-slide-to="2" aria-label="Slide 3"></button>
-           <!--  <button type="button" data-bs-target="#ai-cok-carousel"
-               data-bs-slide-to="3" aria-label="Slide 4"></button> -->
+            <button type="button" data-bs-target="#ai-cok-carousel"
+               data-bs-slide-to="3" aria-label="Slide 4"></button>
 
          </div>
          <!-- AI contents Start -->
@@ -66,9 +66,9 @@
                         </div>
                         <img src="../img/aicc/btn_planner_m_cos1.png" class="mo" alt="">
                      </div>
+                  </div>
                      <img src="../img/aicc/img_robot1.png" alt=""
                         class="d-block mx-auto robot mt-4">
-                  </div>
                </div>
             </div>
             <form action="aiccShow">
@@ -260,7 +260,7 @@
                   </div>
                </div>-->
                <!-- Content3 | 기간 선택 -->
-               <!-- <div class="carousel-item ">
+               <div class="carousel-item ">
                   <div class="ai-box ai-filter-box col-10 col-md-7 text-center">
                      <div class="row f14 mx-0 gap-4">
                         <div class="col-12 mb-5 pb-5">
@@ -293,7 +293,7 @@
                            </div>
                         </div>
                      </div>
-                     다음에 하기 btn
+                     <!-- 다음에 하기 btn -->
                      <div class="py-4 mt-5">
                         <button type="submit"
                            class="btn rounded-pill btn-outline-light next-btn mt-5">여기까지만
@@ -301,7 +301,7 @@
                      </div>
                   </div>
 
-               </div> -->
+               </div>
                <!-- Content4 | 테마 선택 -->
                <div class="carousel-item ">
                   <div class="ai-box ai-filter-box col-10 col-md-7 text-center">
