+++ src/main/resources/templates/thymeleaf/info/Event_view.html
@@ -42,9 +42,9 @@
 		</div>
 		<div class="row px-xl-5 justify-content-center ">
 			<div class="col px-xl-5 my-5" >
-				<h4 class="mt-2 mt-lg-5 mb-3 px-lg-3" data-aos="fade-up">상세 정보</h4>
-				<hr>
-				<p class="px-lg-5 textcontent px-3"
+				<h4 class="mt-2 mt-lg-5 mb-3 px-lg-3 mx-5" data-aos="fade-up">상세 정보</h4>
+				<hr class="mx-5">
+				<p class="px-lg-5 textcontent mx-5"
 					th:text="${view.getEvent_content()}" data-aos="fade-up">내용 들어가는 곳</p>
 
 
@@ -101,29 +101,29 @@
 					</div>
 				</div>
 				<div class="row mb-5 px-md-5 justify-content-center text-center">
-					<div class="col-12 col-lg-8 ">
-						<button class="btn btn-mainbg">
+					<div class="col-12 col-lg-8 my-1">
+						<button class="btn btn-mainbg my-1">
 							# <span th:text="${view.getEvent_tag()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_tag2()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_tag3()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_tag4()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_tag5()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_area()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_area2()}"></span>
 						</button>
-						<button class="btn btn-pink">
+						<button class="btn btn-pink my-1">
 							# <span th:text="${view.getEvent_parking()}"></span>
 						</button>
