+++ src/main/resources/templates/thymeleaf/map/mapDay1.html
@@ -656,13 +656,14 @@
       var mapy24 = /*[[${aiccCafe2.get(0).getEvent_mapY()}]]*/0; // 이벤트의 경도 값
       var mapx25 = /*[[${aiccFood4.get(0).getEvent_mapX()}]]*/0; // 이벤트의 위도 값
       var mapy25 = /*[[${aiccFood4.get(0).getEvent_mapY()}]]*/0; // 이벤트의 경도 값
-      var mapx26 = /*[[${hotel.get(0).getEvent_mapX()}]]*/0; // 이벤트의 위도 값
-      var mapy26 = /*[[${hotel.get(0).getEvent_mapY()}]]*/0; // 이벤트의 경도 값
        var name21 = /*[[${aicc3.get(0).getEvent_name()}]]*/""; // 이벤트의 위도 값
        var name22 = /*[[${aiccFood3.get(0).getEvent_name()}]]*/""; // 이벤트의 경도 값
        var name23 = /*[[${aicc4.get(0).getEvent_name()}]]*/""; // 이벤트의 위도 값
       var name24 = /*[[${aiccCafe2.get(0).getEvent_name()}]]*/""; // 이벤트의 경도 값
        var name25 = /*[[${aiccFood4.get(0).getEvent_name()}]]*/""; // 이벤트의 경도 값
+      
+      var mapx26 = /*[[${hotel.get(0).getEvent_mapX()}]]*/0; // 이벤트의 위도 값
+      var mapy26 = /*[[${hotel.get(0).getEvent_mapY()}]]*/0; // 이벤트의 경도 값
        var name26 = /*[[${hotel.get(0).getEvent_name()}]]*/""; // 이벤트의 경도 값
       // 문자열로 가져온 값을 parseFloat 함수를 사용하여 숫자로 변환
       mapx = parseFloat(mapx);
